Moved step2 to new file, reorganized code
